feat: demo markdown input
